+++ scrapyd/app.py
@@ -8,8 +8,9 @@ from twisted.python import log
 from twisted.cred.portal import Portal
 from twisted.web.guard import HTTPAuthSessionWrapper, BasicCredentialFactory
 
-from .interfaces import IEggStorage, IPoller, ISpiderScheduler, IEnvironment
-from .eggstorage import FilesystemEggStorage
+from scrapy.utils.misc import load_object
+
+from .interfaces import IEggStorage, IJobStorage, IPoller, ISpiderScheduler, IEnvironment
 from .scheduler import SpiderScheduler
 from .poller import QueuePoller
 from .environ import Environment
@@ -41,14 +42,20 @@ def application(config):
     poll_interval = config.getfloat('poll_interval', 5)
 
     poller = QueuePoller(config)
-    eggstorage = FilesystemEggStorage(config)
     scheduler = SpiderScheduler(config)
     environment = Environment(config)
 
     app.setComponent(IPoller, poller)
-    app.setComponent(IEggStorage, eggstorage)
     app.setComponent(ISpiderScheduler, scheduler)
     app.setComponent(IEnvironment, environment)
+
+    jspath = config.get('jobstorage', 'scrapyd.jobstorage.MemoryJobStorage')
+    jscls = load_object(jspath)
+    jobstorage = jscls(config)
+    app.setComponent(IJobStorage, jobstorage)
+    eggstorage = config.get('eggstorage', 'scrapyd.eggstorage.FilesystemEggStorage')
+    eggstoragecls = load_object(eggstorage)
+    app.setComponent(IEggStorage, eggstoragecls(config))
 
     laupath = config.get('launcher', 'scrapyd.launcher.Launcher')
     laucls = load_object(laupath)