+++ scrapyd/interfaces.py
@@ -17,6 +17,9 @@ class IEggStorage(Interface):
         """Return the list of versions which have eggs stored (for the given
         project) in order (the latest version is the currently used)."""
 
+    def list_projects():
+        """Return the list of projects from the stored eggs """
+
     def delete(project, version=None):
         """Delete the egg stored for the given project and version. If should
         also delete the project if no versions are left"""
@@ -107,3 +110,18 @@ class IEnvironment(Interface):
         `message` is the message received from the IPoller.next() method
         `slot` is the Launcher slot where the process will be running.
         """
+
+class IJobStorage(Interface):
+    """A component that handles storing and retrieving finished jobs. """
+
+    def add(job):
+        """Add a finished job in the storage. """
+
+    def list():
+        """Return a list of the finished jobs. """
+
+    def __len__():
+        """Return a number of the finished jobs. """
+
+    def __iter__():
+        """Iterate over the finished jobs. """