+++ scrapyd/website.py
@@ -93,7 +93,10 @@ monitoring)</p>
 </body>
 </html>
 """ % vars
-        return s.encode('utf-8')
+        txrequest.setHeader('Content-Type', 'text/html; charset=utf-8')
+        s = s.encode('utf8')
+        txrequest.setHeader('Content-Length', str(len(s)))
+        return s
 
 
 def microsec_trunc(timelike):
@@ -222,5 +225,6 @@ class Jobs(resource.Resource):
     def render(self, txrequest):
         doc = self.prep_doc()
         txrequest.setHeader('Content-Type', 'text/html; charset=utf-8')
+        doc = doc.encode('utf-8')
         txrequest.setHeader('Content-Length', str(len(doc)))
-        return doc.encode('utf-8')
+        return doc